+++ Engine/Utility/Utilities.cs
@@ -15,17 +15,17 @@ namespace Engine.Utility
 			return Enum.GetValues(typeof(T)).Length;
 		}
 
-		public static ivec2 ComputeOrigin(uint width, uint height, Alignments alignment)
+		public static ivec2 ComputeOrigin(int width, int height, Alignments alignment)
 		{
 			bool left = (alignment & Alignments.Left) > 0;
 			bool right = (alignment & Alignments.Right) > 0;
 			bool top = (alignment & Alignments.Top) > 0;
 			bool bottom = (alignment & Alignments.Bottom) > 0;
 
-			uint x = left ? 0 : (right ? width : width / 2);
-			uint y = top ? 0 : (bottom ? height : height / 2);
+			int x = left ? 0 : (right ? width : width / 2);
+			int y = top ? 0 : (bottom ? height : height / 2);
 
-			return new ivec2((int)x, (int)y);
+			return new ivec2(x, y);
 		}
 	}
 }
